requirements.txt
torch>=1.9.0
torchvision>=0.10.0
transformers>=4.20.0
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
Pillow>=8.3.0
opencv-python>=4.5.0
onnx>=1.12.0
onnxruntime>=1.12.0
torch2onnx
matplotlib>=3.4.0
seaborn>=0.11.0
tqdm>=4.62.0
underthesea>=6.0.0
vncorenlp>=1.0.3

Tải python: https://www.python.org/ftp/python/3.13.3/python-3.13.3-amd64.exe
Bấm windows + R gõ cmd -> enter
Nhập lệnh pip install torch torchvision transformers pandas numpy scikit-learn Pillow opencv-python onnx onnxruntime torch2onnx matplotlib seaborn tqdm underthesea vncorenlp
Chờ nó cài hết là xong
Nhớ cài Visual studio nữa, tải hết các phần của C# là đc